requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
stripe==2.34.0
ulid-py==0.0.9
urllib3==1.25.3
Werkzeug==0.15.5
gunicorn==19.9.0